llama : use double literals for double constants

+++ src/llama-sampler.cpp
@@ -1058,7 +1058,7 @@ static void llama_sampler_dist_apply(struct llama_sampler * smpl, llama_token_da
     }
 
     // apply softmax to obtain the probabilities
-    double sum_cum = 0.0f;
+    double sum_cum = 0.0;
     for (size_t i = 0; i < cur_p->size; ++i) {
         float p = expf(cur_p->data[i].logit - max_l);
         cur_p->data[i].p = p;
@@ -1069,10 +1069,10 @@ static void llama_sampler_dist_apply(struct llama_sampler * smpl, llama_token_da
     // sample from the obtained probabilities and normalize the probs in a single pass
     // this is ~3x faster on Mac with full gpt-oss vocab than the version below
     //
-    std::uniform_real_distribution<double> dist(0.0f, 1.0f);
+    std::uniform_real_distribution<double> dist(0.0, 1.0);
     const double rnd = dist(ctx->rng);
 
-          double sum_run = 0.0f;
+          double sum_run = 0.0;
     const double sum_tgt = sum_cum*rnd;
 
     bool found = false;
@@ -1208,8 +1208,8 @@ static void llama_sampler_dist_backend_set_input(struct llama_sampler * smpl) {
     // std::uniform_real_distribution<double> and
     // std::uniform_real_distribution<float> with same rng will produce
     // different sequences).
-    std::uniform_real_distribution<double> dist(0.0f, 1.0f);
-    const float rnd = dist(sctx->rng);
+    std::uniform_real_distribution<double> dist(0.0, 1.0);
+    const float rnd = static_cast<float>(dist(sctx->rng));
 
     ggml_backend_tensor_set(sctx->inp_uniform, &rnd, 0, sizeof(float));
 }
